Extract engine tests into a shared engine case
